third_party/WebKit/LayoutTests/editing/execCommand/indent-list-item.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable><ul><li>Foo</li><li>|Bar</li><li>Baz</li></ul></div>',
    'indent',
    '<div contenteditable><ul><li>Foo</li><ul><li>|Bar</li></ul><li>Baz</li></ul></div>'),
    'Indent a list item in middle of list');
</script>
